requirement.txt
asgiref==3.6.0
attrs==22.2.0
az-iranian-bank-gateways==1.8.5
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
deploy==1.9.1
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.0.0
ghasedakpack==0.1.13
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
lxml==4.9.2
packaging==23.0
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
pycryptodome==3.17
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
zeep==4.2.1
